API: Rename from_shape_fn_memory_order to from_shape_simple_fn
And let the closure take no argument.

The rationale is that if a counter is needed, it can easily be added by
the caller. It's a duplicate counter then, but with luck it could be
conflated with the other by the compiler?

Use it in Array::default, the other usages are not significant.